billing_api/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-jose==3.3.0
requests==2.28.2
python-dotenv==0.21.1
pydantic==1.10.7